extract addLocalFile as reusable method in VideoService.groovy
